[feat/165] cleaning
